+++ src/ExtendedKey.ts
@@ -18,26 +18,26 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+import * as bip32 from 'bip32';
+import { BIP32 } from 'bip32';
+const bs58check = require('bs58check');
+
 // internal dependencies
 import { 
-    ExtendedKeyNode,
+    CurveAlgorithm,
+    KeyEncoding,
     Network,
+    NodeInterface,
+    NodeEd25519,
 } from '../index';
 
 /**
- * Class `ExtendedKey` describes an extended key as described in 
- * the Bitcoin BIP32 standard which can be found at following
- * URL:
+ * Class `ExtendedKey` describes a hierarchical deterministic extended
+ * key that can be derived (or not if leaf node). This hierarchical 
+ * deterministic key derivation feature is described in the Bitcoin 
+ * BIP32 standard which can be found at following URL:
  *
  *     https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki
- *
- * It will be adapted to work with ED25519 elliptic curve because BIP32
- * is only compatible with secp256k1 elliptic curve.
- *
- * The work on this BIP32-ED25519 will be discussed and described in 
- * the following [NIP](https://github.com/nemtech/NIP/issues/12): 
- *
- *     https://github.com/nemtech/NIP/issues/12
  *
  * This class *uses* features provided by the `bitcoinjs/bip32` package
  * and therefor is licensed under the BSD-2 Clause License as mentioned
@@ -50,32 +50,81 @@ import {
  */
 export class ExtendedKey {
 
-    protected node: ExtendedKeyNode;
-
     /**
      * Construct an `ExtendedKey` object out of its' base58 payload.
-     *
-     * @example Example of `payload` values include:
-     *
-     *   `xpub661MyMwAqRbcFtXgS5sYJABqqG9YLmC4Q1Rdap9gSE8NqtwybGhePY2gZ29ESFjqJoCu1Rupje8YtGqsefD265TMg7usUDFdp6W1EGMcet8`
-     *   `xprv9s21ZrQH143K3QTDL4LXw2F7HEK3wJUD2nW2nRk4stbPy6cq3jPPqjiChkVvvNKmPGJxWUtg6LnF5kejMRNNU3TGtRBeJgk33yuGBxrMPHi`
-     *   `xpub68Gmy5EdvgibQVfPdqkBBCHxA5htiqg55crXYuXoQRKfDBFA1WEjWgP6LHhwBZeNK1VTsfTFUHCdrfp1bgwQ9xv5ski8PX9rL2dZXvgGDnw`
-     *   `xprv9uHRZZhk6KAJC1avXpDAp4MDc3sQKNxDiPvvkX8Br5ngLNv1TxvUxt4cV1rGL5hj6KCesnDYUhd7oWgT11eZG7XnxHrnYeSvkzY7d2bhkJ7`
-     *
-     * @param   base58Payload   {string}
+     * 
+     * @see https://github.com/bitcoinjs/bip32/blob/master/ts-src/bip32.ts
+     * @param   node   {BIP32}
      */
     constructor(/**
-                 * The Base58 payload of the extended key.
-                 * @var {string}
+                 * The hyper-deterministic node.
+                 * @var {BIP32 | NodeEd25519}
                  */
-                public readonly base58Payload: string
+                public readonly node: BIP32 | NodeEd25519,
+                /**
+                 * The hyper-deterministic node network.
+                 * @var {Network}
+                 */
+                public network: Network = Network.BITCOIN
     ) {
-        this.node = ExtendedKeyNode.createFromBase58(base58Payload);
+        if (this.node instanceof NodeEd25519
+            && this.network !== Network.CATAPULT) {
+            throw new Error('Given node is of type NodeEd25519 but network does not match Network.CATAPULT.');
+        }
+        // } else if (this.node instanceof nist256p1) { ..
+        else if (! (this.node instanceof NodeEd25519)
+                 && this.network !== Network.BITCOIN) {
+            throw new Error('Given node is of type BIP32 but network does not match Network.BITCOIN.');
+        }
     }
 
     /**
-     * Create an extended key with the master seed.
+     * Create an extended key hyper-deterministic node by its' Base58
+     * payload.
      *
+     * This method uses the `bitcoinjs/bip32` function named `fromBase58`
+     * and creates an extended key node by parsing the Base58 binary
+     * representation.
+     *
+     * @param payload 
+     */
+    public static createFromBase58(
+        payload: string,
+        network: Network = Network.BITCOIN
+    ): ExtendedKey {
+
+        if (network === Network.CATAPULT) {
+        // use NodeEd25519 node implementation
+
+            // interpret payload
+            const node = NodeEd25519.fromBase58(payload);
+
+            // instanciate our ExtendedKey
+            return new ExtendedKey(node, network);
+        }
+        // else {
+        // use BIP32 node implementation
+
+        // interpret payload
+        const node = bip32.fromBase58(payload);
+
+        // instanciate our ExtendedKey
+        return new ExtendedKey(node, network);
+    }
+
+    /**
+     * Create an extended key hyper-deterministic node with the master
+     * seed.
+     *
+     * This method uses the `bitcoinjs/bip32` function named `fromSeed`
+     * and creates an extended key node by creating HMAC-SHA512 hash
+     * of the words 'Bitcoin seed' appended with the `seed` binary
+     * representation.
+     *
+     * The result is split in 2 parts where the left most 32 bytes are
+     * the private and right most 32 bytes are the public key.
+     *
+     * @see https://github.com/bitcoinjs/bip32/blob/master/src/bip32.js#L265
      * @param   seed    {string}
      * @param   network {Network}
      * @return  {ExtendedKey}
@@ -84,44 +133,180 @@ export class ExtendedKey {
         seed: string,
         network: Network = Network.BITCOIN
     ): ExtendedKey {
-        const node = ExtendedKeyNode.createFromSeed(seed, network);
-        return new ExtendedKey(node.toBase58());
+
+        if (network === Network.CATAPULT) {
+        // use NodeEd25519 node implementation
+
+            // use hexadecimal seed
+            const node = NodeEd25519.fromSeed(Buffer.from(seed, 'hex'));
+
+            // instanciate our ExtendedKey
+            return new ExtendedKey(node, network);
+        }
+        // else {
+        // use BIP32 node implementation
+
+        // use hexadecimal seed
+        const node = bip32.fromSeed(Buffer.from(seed, 'hex'));
+
+        // instanciate our ExtendedKey
+        return new ExtendedKey(node, network);
     }
 
     /**
-     * Create a child private key from an extended key.
+     * Derive hyper-deterministic node by `path`.
+     *
+     * Default account layer should derive path `m/44'/43'/0'/0/0`.
+     *
+     * @see https://github.com/nemtech/NIP/issues/12
+     * @param path 
+     */
+    public derivePath(
+        path: string
+    ): ExtendedKey {
+
+        // derive path with specialized `derivePath`
+        const derived = this.node.derivePath(path);
+
+        if (derived instanceof NodeEd25519) {
+        // use NodeEd25519 node implementation
+            return new ExtendedKey(derived as NodeEd25519, this.network);
+        }
+        // else {
+        // use BIP32 node implementation
+
+        return new ExtendedKey(derived as BIP32, this.network);
+    }
+
+    /**
+     * Return whether an extended key node is neutered or not.
+     *
+     * Neutered = Public Key only
+     * Not Neutered = Private Key available
+     *
+     * @return {boolean}
+     */
+    public isNeutered(): boolean {
+        // forward to `bitcoinjs/bip32`
+        return this.node.isNeutered();
+    }
+
+    /**
+     * Return whether the current `node` is a master key node or not.
+     *
+     * @return {boolean}
+     */
+    public isMaster(): boolean {
+        //XXX read parentFingerprint instead of decode
+        const base58 = this.node.toBase58();
+        const buffer = bs58check.decode(base58);
+        const parent = buffer.readUInt32BE(5);
+
+        return parent === 0x00000000;
+    }
+
+    /**
+     * Get a neutered hyper-deterministic node. This corresponds to
+     * a public key only extended key.
      * 
-     * This is possible only with **not-neutered** extended key nodes.
+     * From a neutered HD-node, users can only generate **public child
+     * keys** and no **private child keys**.
+     *
+     * @see https://github.com/bitcoinjs/bip32/blob/master/ts-src/bip32.ts#L118
+     * @return {ExtendedKey}    The neutered HD-node
+     */
+    public getPublicNode(): ExtendedKey {
+
+        // create new node from neutered
+        const node = this.node.neutered();
+
+        if (node instanceof NodeEd25519) {
+        // use NodeEd25519 node implementation
+
+            return new ExtendedKey(node as NodeEd25519, this.network);
+        }
+        // else {
+        // use BIP32 node implementation
+
+        return new ExtendedKey(node as BIP32, this.network);
+    }
+
+    /**
+     * This method proxies the conversion to base58 format
+     * to the `bitcoinjs/bip32` library.
+     *
+     * @return {string}
+     */
+    public toBase58(): string {
+        // forward to `bitcoinjs/bip32`
+        return this.node.toBase58();
+    }
+
+    /**
+     * Get the private key of the HD-node.
+     *
+     * This method defaults to returning the hexadecimal notation of
+     * the key. Use `KeyEncoding.ENC_BIN` if you need the binary form.
      * 
-     * @param   derivationPath  {string}    Example: "m/44'/43'/0'/0/0"
-     * @return  {ExtendedKey}
+     * @see {KeyEncoding}
+     * @return  {string}
      * @throws  {Error}     On use of this method with neutered extended keys (public keys).
      */
-    getChildPrivateKey(
-        derivationPath: string
-    ): ExtendedKey {
-        if (this.node.isNeutered()) {
-            throw new Error('Cannot derive child private key out of extended public key.');
+    public getPrivateKey(
+        encoding: KeyEncoding = KeyEncoding.ENC_HEX
+    ): string | Buffer {
+        if (this.isNeutered()) {
+            throw new Error('Cannot read private key out of extended public key.');
         }
 
-        const child = this.node.derivePath(derivationPath);
-        return new ExtendedKey(child.toBase58());
+        // return encoded private key (default hexadecimal format)
+        return this.encodeAs(this.node.privateKey, encoding);
     }
 
     /**
-     * Create a child public key from an extended key.
-     * 
-     * This is possible with either of neutered and non-neutered extended
-     * key nodes.
+     * Get the public key in hexadecimal notation.
      *
-     * @param   derivationPath  {string}    Example: "m/44'/43'/0'/0/0"
-     * @return  {ExtendedKey}
+     * This method defaults to returning the hexadecimal notation of
+     * the key. Use `KeyEncoding.ENC_BIN` if you need the binary form.
+     *
+     * @see {KeyEncoding}
+     * @return  {string}
+     * @throws  {Error}     On use of this method with neutered extended keys (public keys).
      */
-    getChildPublicKey(
-        derivationPath: string
-    ): ExtendedKey {
-        const child = this.node.derivePath(derivationPath);
-        return new ExtendedKey(child.toBase58());
+    public getPublicKey(
+        encoding: KeyEncoding = KeyEncoding.ENC_HEX
+    ): string | Buffer {
+
+        // @see https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki
+        // ser-p(P) serializes the coordinate and prepends either 0x02 or 0x03 to it.
+        // drop first byte for 32-bytes public key
+        let publicKey = this.node.publicKey;
+        if (this.node.publicKey.byteLength === 33) {
+            publicKey = this.node.publicKey.slice(1);
+        }
+
+        // return encoded public key (default hexadecimal format)
+        return this.encodeAs(publicKey, encoding);
+    }
+
+    /**
+     * Encode a key into `encoding`. Default `encoding` is `KeyEncoding.ENC_HEX`
+     * which results in a hexadecimal notation of the key.
+     *
+     * @param key 
+     * @param encoding 
+     */
+    protected encodeAs(
+        key: Buffer,
+        encoding: KeyEncoding = KeyEncoding.ENC_HEX
+    ): string | Buffer {
+        if (encoding === KeyEncoding.ENC_HEX) {
+            // return hexadecimal notation
+            return key.toString('hex');
+        }
+
+        // return binary Buffer
+        return key;
     }
 
 }